+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## Unreleased
 
+### Fixed
+
+- Undo a hardcoded directory string replace to support multiple module directories
+
 ## 2.0.0 (2019-01-21)
 
 ### Added